SKETCH-2618: Refactoring DrawMonomerSceneTool to create AbstractMonomerSceneTool
